fix: point integration tests at the current development branch for istio charms
Previously, integration tests deployed istio-k8s from channel `latest/edge` because that was the leading edge development branch.  Now that we've cut a `1` track for the prior release and a `2` track for current development, we need to point at `2/edge`.  We cannot simply point at `edge`, because `1` is the default track
